aievent/requirements.txt
firebase-functions~=0.1.0
google-generativeai
functions-framework
requests
firebase-admin